% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session2CO.R
\name{session2CO}
\alias{session2CO}
\title{Convert sessionInfo to Code Ocean}
\usage{
session2CO(sess = NULL, path = paste0(getwd(), "/session2CO.txt"))
}
\arguments{
\item{sess}{Stored sessionInfo}

\item{path}{A path and filename for to store the reuslts}
}
\value{
Writes a file with RUN Rscript lines per package
}
\description{
Converts the output of sessionInfo() in lines of RUN Rscript code for Code Ocean
}
\examples{
sess <- sessionInfo()
session2CO(sess = sess, path = "session2CO.txt")
}
\keyword{Code}
\keyword{Docker}
\keyword{Ocean,}
\keyword{sessionInfo,}
